Moving 'src/public/js/game.js' to 'src/public/js/managers/game.manager.js'
