+++ src/public/js/managers/game.manager.js
